LayoutTests/platform/glib/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-mixed-content-to-outscope.https-expected.txt

CONSOLE MESSAGE: The page at https://web-platform.test:9443/service-workers/service-worker/resources/fetch-mixed-content-iframe-inscope-to-outscope.html was allowed to display insecure content from http://web-platform.test:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load http://web-platform.test:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.
CONSOLE MESSAGE: The page at https://web-platform.test:9443/service-workers/service-worker/resources/fetch-mixed-content-iframe-inscope-to-outscope.html was allowed to display insecure content from http://www1.web-platform.test:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load http://www1.web-platform.test:8800/service-workers/service-worker/resources/fetch-access-control.py?PNGIMAGE.


PASS Verify Mixed content of fetch() in a Service Worker
